pyproject.toml
[tool.poetry]
name = "labster"
version = "2.1.10"
description = ""
authors = ["Abilian SAS"]

include = [
    "poetry.lock",
    "labster/static/**/*",
    "migrations/**/*",
    "etc/**/*",
    "boot/**/*",
    "scripts/**/*",
    #    "*.py",
]


[tool.poetry.dependencies]
python = "^3.6"

abilian-core = ">=0.11"
psycopg2-binary = "*"
Flask-Redis = "*"

# Pinned for now
pandas = "<0.25"

# Seems tricky
cairocffi = "^1.0.2"
Flask-WeasyPrint = "*"

# Swagger API
# apispec = "<0.39"
# flask-apispec = "==0.4.2"

# Marshmallow & friends
marshmallow = "*"
marshmallow-sqlalchemy = "*"
marshmallow-annotations = "*"
flask-resty = "^0.21.4"
flask-smorest = "*"

# Used when sending emails
html2text = "*"

# Used by workflow
inflection = "*"

# Used for parsing LDIF
python-ldap = ">=3"

# Used by CAS auth
requests = { version = "^2.13", extras = [ "security" ] }

# Used for data classes
attrs = "*"

# Functional libraries (we should stanfdardize on one !).
toolz = "*"
ramda = "^0.5.2"

# Other
unidecode = "*"
dotted = "*"
iso8601 = "*"
python-dotenv = "^0.10.1"
flask-profiler = "^1.8"
openpyxl = "*"

structlog = "^19.2"
nox = "^2019.8"
jsonrpcserver = "^4.0"
webargs = "^5.5"
glom = "^19.2"

xlrd = "^1.2"
flask-injector = { git = "https://github.com/abilian/flask_injector.git" }

# q = "*"

#: Production
gunicorn = "*"
raven = "*"

# supervisor for py3 isn't released yet.
# It isn't a pb to run the app though.
# supervisor >= 4
# git+https://github.com/Supervisor/supervisor


[tool.poetry.dev-dependencies]
flake8 = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
flake8-mutable = "*"
flake8-super-call = "*"
flake8-pytest = "*"
flake8-tidy-imports = "*"
mypy = "*"
mastool = "*"
pylint = "*"

dephell = "^0.7.4"
# Temp (otherwise it resolves to 0.0.1 !)
bowler = "^0.8.0"
fissix = "19.2b1"

pre-commit = "*"

pytest = "*"
pytest-cov = "*"
pytest-mock = "*"
tox = "*"
typeguard = "*"

Flask-Linktester = "*"

honcho = "*"

# Formatters
isort = "*"
black = { version = "*", allows-prereleases = true }
tabulate = "^0.8.3"
gitchangelog = "^3.0"
pytest-randomly = "^3.1"


[tool.poetry.scripts]
labster = "labster.cli:cli"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

# [build-system]
# requires = ["poetry>=0.12"]
# build-backend = "poetry.masonry.api"
